Add LoadingScreen to init app.
