+++ lib/widgets/loading_screen/loading_screen.dart
@@ -1,0 +1,54 @@
+import 'package:adventskalender/di_container.dart';
+import 'package:adventskalender/widgets/home_screen/home_screen.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:hive/hive.dart';
+import 'package:path_provider/path_provider.dart';
+
+class LoadingScreen extends StatefulWidget {
+  const LoadingScreen({Key key}) : super(key: key);
+
+  @override
+  _LoadingScreenState createState() => _LoadingScreenState();
+}
+
+class _LoadingScreenState extends State<LoadingScreen> {
+  Future<bool> _initFuture;
+
+  @override
+  void initState() {
+    super.initState();
+
+    _initFuture = _init();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: FutureBuilder(
+        future: _initFuture,
+        builder: (_, snapshot) {
+          if (snapshot.connectionState == ConnectionState.done && snapshot.data) {
+            return HomeScreen();
+          }
+
+          return Container();
+        },
+      ),
+    );
+  }
+
+  Future<bool> _init() async {
+    final dir = await getApplicationDocumentsDirectory();
+    Hive.init(dir.path);
+
+    final db = context.read(databaseServiceProvider);
+    await db.initialize(dir.path);
+
+    final _notificationService = context.read(localNotificationServiceProvider);
+    await _notificationService.initialize();
+    await _notificationService.requestPermissions();
+
+    return true;
+  }
+}
