+++ lib/widgets/my_app.dart
@@ -1,5 +1,5 @@
 import 'package:adventskalender/localizations.dart';
-import 'package:adventskalender/widgets/home_screen/home_screen.dart';
+import 'package:adventskalender/widgets/loading_screen/loading_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 
@@ -13,7 +13,7 @@ class MyApp extends StatelessWidget {
         GlobalWidgetsLocalizations.delegate,
       ],
       supportedLocales: AppLocalizationsDelegate.supportedLocals,
-      home: HomeScreen(),
+      home: LoadingScreen(),
     );
   }
 }
